refactor: :recycle: разделить правила и типы дочерних элементов
Построители правил и структура YAML вынесены в нейтральные модули с реэкспортом по старому пути. Context helper использует точечный импорт типа вместо общего orchestration-модуля.
