+++ .dependency-cruiser-cycle-baseline.json
@@ -91,44 +91,25 @@
     },
     {
       "modules": [
-        "packages/core/metadata/forms/commonObjects/childItems/types.ts",
-        "packages/core/metadata/forms/elements/autoCommandBar/rules.ts",
-        "packages/core/metadata/forms/elements/autoCommandBar/types.ts",
-        "packages/core/metadata/forms/elements/buttonGroup/rules.ts",
-        "packages/core/metadata/forms/elements/buttonGroup/types.ts",
-        "packages/core/metadata/forms/elements/columnGroup/rules.ts",
-        "packages/core/metadata/forms/elements/columnGroup/types.ts",
-        "packages/core/metadata/forms/elements/commandBar/rules.ts",
-        "packages/core/metadata/forms/elements/commandBar/types.ts",
-        "packages/core/metadata/forms/elements/contextMenu/rules.ts",
-        "packages/core/metadata/forms/elements/contextMenu/types.ts",
-        "packages/core/metadata/forms/elements/page/rules.ts",
-        "packages/core/metadata/forms/elements/page/types.ts",
-        "packages/core/metadata/forms/elements/pages/rules.ts",
-        "packages/core/metadata/forms/elements/pages/types.ts",
-        "packages/core/metadata/forms/elements/popup/rules.ts",
-        "packages/core/metadata/forms/elements/popup/types.ts",
-        "packages/core/metadata/forms/elements/searchControlAddition/types.ts",
-        "packages/core/metadata/forms/elements/searchStringAddition/types.ts",
-        "packages/core/metadata/forms/elements/usualGroup/rules.ts",
-        "packages/core/metadata/forms/elements/usualGroup/types.ts",
-        "packages/core/metadata/forms/elements/viewStatusAddition/types.ts"
+        "packages/core/metadata/orchestration/metadataCollection/fromYAMLToXML.ts",
+        "packages/core/metadata/orchestration/metadataItem/fromYAMLToXML.ts",
+        "packages/core/metadata/orchestration/property/fromYAMLToXML.ts"
       ],
-      "dependencyCount": 35
+      "dependencyCount": 4
     },
     {
       "modules": [
-        "packages/core/metadata/context/helpers.ts",
         "packages/core/metadata/orchestration/formElement/ruleFactory.ts",
-        "packages/core/metadata/orchestration/formElement/toJSONSchema.ts",
+        "packages/core/metadata/orchestration/formElement/toJSONSchema.ts"
+      ],
+      "dependencyCount": 2
+    },
+    {
+      "modules": [
         "packages/core/metadata/orchestration/index.ts",
-        "packages/core/metadata/orchestration/metadataCollection/fromYAMLToXML.ts",
-        "packages/core/metadata/orchestration/metadataCollection/index.ts",
-        "packages/core/metadata/orchestration/metadataItem/fromYAMLToXML.ts",
-        "packages/core/metadata/orchestration/property/fromYAMLToXML.ts",
         "packages/core/metadata/orchestration/property/toEnterprise.ts"
       ],
-      "dependencyCount": 17
+      "dependencyCount": 2
     }
   ]
 }
