+++ packages/core/metadata/forms/clientApplicationForm/rules.ts
@@ -4,7 +4,7 @@ import { stringOrNumberRule } from "../../commonObjects/stringOrNumber/types"
 import { usePurposesRule } from "../../commonObjects/usePurposes/types"
 import { choiceParametersRule } from "../../commonObjects/\u0441hoiceParameters/types"
 import { autoCommandBarRule, conditionalAppearanceRule, externalFormItemFileRule } from "./builders"
-import { groupChildItemsRule } from "../commonObjects/childItems/types"
+import { groupChildItemsRule } from "../commonObjects/childItems/rules"
 import { commandInterfaceRule } from "../commonObjects/commandInterface/types"
 import { commandSetRule } from "../commonObjects/commandSet/types"
 import { eventsRule } from "../commonObjects/event/types"
