+++ packages/core/metadata/forms/commonObjects/childItems/contracts.ts
@@ -1,0 +1,11 @@
+import type {
+  CollectableElementToYAML,
+  CollectableElementType,
+} from "../../../orchestration/formElement/types"
+
+export type FormElementTreeYAML = Record<string, FormElementTreeNodeYAML>
+
+export type FormElementTreeNodeYAML = {
+  Вид: CollectableElementToYAML<CollectableElementType>
+  Элементы?: FormElementTreeYAML
+} & Record<string, unknown>
