+++ packages/core/metadata/forms/elements/contextMenu/types.ts
@@ -2,7 +2,7 @@ import { StringboolYAML } from "../../../commonObjects/boolean/types"
 import { FormTypeByRule } from "../../../orchestration/metadataItem/element"
 import { EnterpriseType } from "../../../orchestration/metadataItem/enterprise"
 import * as SE from "../../../systemEnumerations/types"
-import { FormElementTreeYAML } from "../../commonObjects/childItems/types"
+import type { FormElementTreeYAML } from "../../commonObjects/childItems/contracts"
 import { ContextMenuRules } from "./rules"
 
 export type ContextMenu = FormTypeByRule<typeof ContextMenuRules>
