+++ packages/core/metadata/forms/elements/pages/rules.ts
@@ -1,5 +1,5 @@
 import { associatedTableRule } from "../../../commonObjects/metadataValue/types"
-import { pagesChildItemsRule } from "../../commonObjects/childItems/types"
+import { pagesChildItemsRule } from "../../commonObjects/childItems/rules"
 import { eventsRule } from "../../commonObjects/event/types"
 import { stringRule } from "../../../commonObjects/string/types"
 import { systemEnumerationRule } from "../../../systemEnumerations/types"
