+++ packages/core/metadata/forms/elements/popup/rules.ts
@@ -1,6 +1,6 @@
 import { colorRule } from "../../../commonObjects/color/types"
 import { pictureRule } from "../../../commonObjects/picture/types"
-import { commandBarChildItemsRule } from "../../commonObjects/childItems/types"
+import { commandBarChildItemsRule } from "../../commonObjects/childItems/rules"
 import { stringRule } from "../../../commonObjects/string/types"
 import { systemEnumerationRule } from "../../../systemEnumerations/types"
 import { registerElementRule } from "../../../orchestration/formElement/ruleFactory"
