+++ packages/core/metadata/forms/elements/searchControlAddition/types.ts
@@ -7,7 +7,7 @@ import { UserVisibleYAML } from "../../../commonObjects/userVisible/types"
 import { FormTypeByRule } from "../../../orchestration/metadataItem/element"
 import { EnterpriseType } from "../../../orchestration/metadataItem/enterprise"
 import * as SE from "../../../systemEnumerations/types"
-import { FormElementTreeYAML } from "../../commonObjects/childItems/types"
+import type { FormElementTreeYAML } from "../../commonObjects/childItems/contracts"
 import { ContextMenuYAML } from "../contextMenu/types"
 import { ExtendedTooltipYAML } from "../extendedTooltip/types"
 import { SearchControlAdditionRules, SingleSearchControlAdditionRules } from "./rules"
